Break up Semantic System section into parts
