+++ src/stories/FlexWrap.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const FlexWrap = styled.div(({ theme }) => ({
+  display: 'flex',
+  flexWrap: 'wrap',
+  gap: theme.spacing.large,
+}))
