+++ src/stories/_SemanticSystem.stories.tsx
@@ -1,9 +1,17 @@
 import styled, { useTheme } from 'styled-components'
 
 import { type FillLevel } from '../components/contexts/FillLevelContext'
-import { type styledTheme } from '..'
 import Divider from '../components/Divider'
 import { baseSpacing } from '../theme/spacing'
+
+import { ItemLabel } from './ItemLabel'
+import { FilledBox } from './FilledBox'
+import { FlexWrap } from './FlexWrap'
+
+export default {
+  title: 'Semantic System',
+  component: null,
+}
 
 const fillLevelToBGColor: Record<FillLevel, string> = {
   0: 'fill-zero',
@@ -12,37 +20,16 @@ const fillLevelToBGColor: Record<FillLevel, string> = {
   3: 'fill-three',
 }
 
-export default {
-  title: 'Semantic System',
-  component: null,
-}
-
-const ItemLabel = styled.div(({ theme }) => ({
-  ...theme.partials.text.caption,
-  marginTop: theme.spacing.xxsmall,
-}))
-
 const BlockWrapper = styled.div(({ theme }) => ({
   marginBottom: theme.spacing.large,
-}))
-
-const FilledBox = styled.div<{ $bgColor?: string }>(({ theme, $bgColor }) => ({
-  width: '64px',
-  height: '64px',
-  backgroundColor: $bgColor || theme.colors['fill-one'],
 }))
 
 function Template({ exampleText }: { exampleText?: string }) {
   return (
     <>
       <Divider
-        text="Colors"
-        marginVertical="xxlarge"
-      />
-      <Colors />
-      <Divider
         text="Shadows"
-        marginVertical="xxlarge"
+        marginBottom="xxlarge"
       />
       <Shadows />
       <Divider
@@ -65,59 +52,7 @@ function Template({ exampleText }: { exampleText?: string }) {
         marginVertical="xxlarge"
       />
       <Spacing />
-      <Divider
-        text="Typography"
-        marginVertical="xxlarge"
-      />
-      <Typography exampleText={exampleText} />
-      <Divider
-        text="Marketing Typography"
-        marginVertical="xxlarge"
-      />
-      <MarketingTypography exampleText={exampleText} />
     </>
-  )
-}
-
-const ColorBox = styled(FilledBox)<{ color: string }>(({ theme, color }) => ({
-  boxShadow: theme.boxShadows.moderate,
-  backgroundColor: (theme.colors as any)[color],
-}))
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const ColorBoxWrap = styled.div((_) => ({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  width: '64px',
-}))
-
-const FlexWrap = styled.div(({ theme }) => ({
-  display: 'flex',
-  flexWrap: 'wrap',
-  gap: theme.spacing.large,
-}))
-
-function Colors() {
-  const theme = useTheme()
-
-  const colors = { ...theme.colors }
-
-  delete colors.blue
-  delete colors.grey
-  delete colors.green
-  delete colors.yellow
-  delete colors.red
-
-  return (
-    <FlexWrap>
-      {Object.entries(colors).map(([key]) => (
-        <ColorBoxWrap key={key}>
-          <ColorBox color={`${key}`} />
-          <ItemLabel>{key}</ItemLabel>
-        </ColorBoxWrap>
-      ))}
-    </FlexWrap>
   )
 }
 
@@ -288,127 +223,10 @@ function Spacing() {
   )
 }
 
-const SemanticText = styled.div<{
-  $typeStyle?: keyof typeof styledTheme.partials.text
-}>(({ theme, $typeStyle: typeStyle }) => ({
-  ...theme.partials.text[typeStyle],
-  marginBottom: theme.spacing.large,
-}))
-
-function Typography({
-  exampleText: txt = 'Lorem ipsum dolor sit amet',
-}: {
-  exampleText: string
-}) {
-  return (
-    <>
-      <SemanticText $typeStyle="h1">H1 - {txt}</SemanticText>
-      <SemanticText $typeStyle="h2">H2 - {txt}</SemanticText>
-      <SemanticText $typeStyle="h3">H3 - {txt}</SemanticText>
-      <SemanticText $typeStyle="h4">H4 - {txt}</SemanticText>
-      <SemanticText $typeStyle="title1">Title 1 - {txt}</SemanticText>
-      <SemanticText $typeStyle="title2">Title 2 - {txt}</SemanticText>
-      <SemanticText $typeStyle="subtitle1">Subtitle 1 - {txt}</SemanticText>
-      <SemanticText $typeStyle="subtitle2">Subtitle 2 - {txt}</SemanticText>
-      <SemanticText $typeStyle="body1Bold">Body 1 (Bold) - {txt}</SemanticText>
-      <SemanticText $typeStyle="body1">Body 1 - {txt}</SemanticText>
-      <SemanticText $typeStyle="body2Bold">Body 2 (Bold) - {txt}</SemanticText>
-      <SemanticText $typeStyle="body2">Body 2 - {txt}</SemanticText>
-      <SemanticText $typeStyle="body2LooseLineHeight">
-        Body 2 Loose Line Height - {txt}
-      </SemanticText>
-      <SemanticText $typeStyle="caption">Caption - {txt}</SemanticText>
-      <SemanticText $typeStyle="badgeLabel">Badge Label - {txt}</SemanticText>
-      <SemanticText $typeStyle="buttonLarge">Large Button - {txt}</SemanticText>
-      <SemanticText $typeStyle="buttonSmall">Small Button - {txt}</SemanticText>
-      <SemanticText $typeStyle="overline">Overline - {txt}</SemanticText>
-      <SemanticText $typeStyle="code">Code - {txt}</SemanticText>
-    </>
-  )
-}
-
-const MktgText = styled.div<{
-  $typeStyle?: keyof typeof styledTheme.partials.marketingText
-}>(({ theme, $typeStyle: typeStyle }) => ({
-  ...theme.partials.marketingText[typeStyle],
-  display: 'block',
-  marginBottom: theme.spacing.large,
-}))
-
-const MarketingInlineLink = styled.a(({ theme }) => ({
-  ...theme.partials.marketingText.inlineLink,
-}))
-
-function MarketingTypography({
-  exampleText: txt = 'Lorem ipsum dolor sit amet',
-}: {
-  exampleText: string
-}) {
-  return (
-    <>
-      <MktgText $typeStyle="bigHeader">
-        Big Header (
-        <MarketingInlineLink href="#">Inline link</MarketingInlineLink>) - {txt}
-      </MktgText>
-      <MktgText $typeStyle="hero1">
-        Hero 1 (<MarketingInlineLink href="#">Inline link</MarketingInlineLink>)
-        - {txt}
-      </MktgText>
-      <MktgText $typeStyle="hero2">
-        Hero 2 (<MarketingInlineLink href="#">Inline link</MarketingInlineLink>)
-        - {txt}
-      </MktgText>
-      <MktgText $typeStyle="title1">
-        Title 1 (<MarketingInlineLink href="#">Inline link</MarketingInlineLink>
-        ) - {txt}
-      </MktgText>
-      <MktgText $typeStyle="title2">
-        Title 2 (<MarketingInlineLink href="#">Inline link</MarketingInlineLink>
-        ) - {txt}
-      </MktgText>
-      <MktgText $typeStyle="subtitle1">
-        Subtitle 1 (
-        <MarketingInlineLink href="#">Inline link</MarketingInlineLink>) - {txt}
-      </MktgText>
-      <MktgText $typeStyle="subtitle2">
-        Subtitle 2 (
-        <MarketingInlineLink href="#">Inline link</MarketingInlineLink>) - {txt}
-      </MktgText>
-      <MktgText $typeStyle="body1Bold">
-        Body 1 (Bold) (
-        <MarketingInlineLink href="#">Inline link</MarketingInlineLink>) - {txt}
-      </MktgText>
-      <MktgText $typeStyle="body1">
-        Body 1 (<MarketingInlineLink href="#">Inline link</MarketingInlineLink>)
-        - {txt}
-      </MktgText>
-      <MktgText $typeStyle="body2Bold">
-        Body 2 (Bold) (
-        <MarketingInlineLink href="#">Inline link</MarketingInlineLink>) - {txt}
-      </MktgText>
-      <MktgText $typeStyle="body2">
-        Body 2 (<MarketingInlineLink href="#">Inline link</MarketingInlineLink>)
-        - {txt}
-      </MktgText>
-      <MktgText $typeStyle="standaloneLink">Standalone link - {txt}</MktgText>
-      <MktgText $typeStyle="componentText">
-        Component text (
-        <MarketingInlineLink href="#">Inline link</MarketingInlineLink>) - {txt}
-      </MktgText>
-      <MktgText $typeStyle="componentLink">Component link - {txt}</MktgText>
-      <MktgText $typeStyle="componentLinkSmall">
-        Small component link - {txt}
-      </MktgText>
-      <MktgText $typeStyle="label">
-        Label (<MarketingInlineLink href="#">Inline link</MarketingInlineLink>)
-        - {txt}
-      </MktgText>
-      <MktgText $typeStyle="navLink">Nav link - {txt}</MktgText>
-    </>
-  )
-}
-
-export const SemanticSystem = Template.bind({})
-SemanticSystem.args = {
+export const Miscellaneous = Template.bind({})
+Miscellaneous.args = {
   exampleText: 'Lorem ipsum dolor sit amet',
 }
+
+export { default as Colors } from './Colors'
+export { default as Typography } from './Typography'
